Restructured PoseidonBatchMerkleTree
