+++ primitives/src/crh/poseidon/parameters.rs
@@ -6,6 +6,7 @@ use algebra::fields::mnt4753::Fr as MNT4753Fr;
 use algebra::biginteger::BigInteger768;
 use algebra::field_new;
 
+#[derive(Clone)]
 pub struct MNT4753PoseidonParameters;
 
 impl FieldBasedHashParameters for MNT4753PoseidonParameters {
@@ -263,8 +264,8 @@ impl PoseidonParameters for MNT4753PoseidonParameters {
         ];
 }
 
+#[derive(Clone)]
 pub struct MNT6753PoseidonParameters;
-
 
 impl FieldBasedHashParameters for MNT6753PoseidonParameters {
     type Fr = MNT6753Fr;
